requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langfuse==2.10.0
openai==1.6.1
tiktoken>=0.5.2
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.28.0
faiss-cpu==1.13.0
pypdf==3.17.4
